Add new helper package
